Cache tar downloads, fix building from mono-repo tar
	- Previously, when a release tar would be used, the file would
	be downloaded even if it was in the cache
	- Don't put the cached files in a dot file. Took me a while to
	find the damn things.
	- Tar release files are now pulled from the github releases

+++ src/resource.rs
@@ -181,13 +181,20 @@ impl Resource {
             }
             Resource::Tar { url } => {
                 info!("Download Tar file: {}", url);
-                let working = cache_dir()?.join(".tar_download");
+                let working = cache_dir()?.join("cached_tar_downloads");
                 fs::create_dir_all(&working)?;
                 let filename = get_filename_from_url(url)?;
                 let path = working.join(&filename);
-                let mut req = reqwest::get(url)?;
-                let mut f = fs::File::create(&path)?;
-                req.copy_to(&mut f)?;
+                if !path.exists() {
+                    let mut req = reqwest::get(url)?;
+                    let mut f = fs::File::create(&path)?;
+                    req.copy_to(&mut f)?;
+                } else {
+                    info!(
+                        "A cached tar file exists at {}, not downloading...",
+                        path.display()
+                    );
+                }
                 Command::new("tar")
                     .arg("xf")
                     .arg(filename)
@@ -198,15 +205,7 @@ impl Resource {
                     .filter(|d| d.file_type().unwrap().is_dir())
                     .nth(0)
                     .expect("Archive does not contains file");
-                for contents in fs::read_dir(d.path())? {
-                    let path = contents?.path();
-                    if path.is_dir() {
-                        let opt = fs_extra::dir::CopyOptions::new();
-                        fs_extra::dir::copy(path, dest, &opt)?;
-                    } else {
-                        fs::copy(&path, dest.join(path.file_name().unwrap()))?;
-                    }
-                }
+                std::fs::rename(d.path(), dest)?;
             }
         }
         Ok(())
